requirements.txt
streamlit
torch
torchvision
transformers
opencv-python
pillow
numpy
pandas
matplotlib
scikit-learn
timm